Move and rename EntityToTreeMapper to mapper folder
